src/requirements.txt
azure-identity==1.23.1
azure-mgmt-network==29.0.0